🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Add status filtering to workflow runs UI with `StatusFilterDropdown` in `WorkflowPage`, `Workflows`, and `TaskHistory`.
>
>   - **Behavior**:
>     - Adds `StatusFilterDropdown` to `WorkflowPage`, `Workflows`, and `TaskHistory` to filter workflow runs by status.
>     - Updates query keys in `useQuery` to include `statusFilters` in `WorkflowPage`, `Workflows`, and `TaskHistory`.
>     - Allows overriding default status options in `StatusFilterDropdown`.
>   - **Components**:
>     - Refactors `StatusFilterDropdown` to use default status options if none are provided.
>     - Removes redundant `statusDropdownItems` from `TaskHistory`.
>   - **Misc**:
>     - Updates `WorkflowPage` and `Workflows` to include status filters in API requests.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for f31d750507db980b205eedc27e271cafd9ea7456. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
